+++ 10xGitHubPolicies.App/Services/GitHub/IGitHubService.cs
@@ -13,6 +13,7 @@ public interface IGitHubService
     Task<Issue> CreateIssueAsync(long repositoryId, string title, string body, IEnumerable<string> labels);
     Task ArchiveRepositoryAsync(long repositoryId);
     Task<bool> IsUserMemberOfTeamAsync(string userAccessToken, string org, string teamSlug);
-    Task<string> GetFileContentAsync(string repoName, string path);
-    Task<string> GetWorkflowPermissionsAsync(long repositoryId);
+    Task<string?> GetFileContentAsync(string repoName, string path);
+    Task<string?> GetWorkflowPermissionsAsync(long repositoryId);
+    Task<IReadOnlyList<Issue>> GetOpenIssuesAsync(long repositoryId, string label);
 }