+++ 10xGitHubPolicies.App/Services/Scanning/ScanningService.cs
@@ -1,11 +1,13 @@
 using _10xGitHubPolicies.App.Data;
 using _10xGitHubPolicies.App.Data.Entities;
+using _10xGitHubPolicies.App.Services.Action;
 using _10xGitHubPolicies.App.Services.Configuration;
 using _10xGitHubPolicies.App.Services.Configuration.Models;
 using _10xGitHubPolicies.App.Services.GitHub;
 using _10xGitHubPolicies.App.Services.Policies;
-using _10xGitHubPolicies.App.Services.Action;
+
 using Hangfire;
+
 using Microsoft.EntityFrameworkCore;
 
 namespace _10xGitHubPolicies.App.Services.Scanning;